requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
blinker==1.8.2
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.14.6
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.3.0
cryptography==39.0.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
editdistance==0.8.1
exceptiongroup==1.2.0
executing==0.8.3
fastapi==0.115.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.66.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
huggingface-hub==0.25.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==7.0.1
importlib_resources==6.4.5
ipykernel==6.28.0
ipython==8.15.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kmodes==0.12.2
Levenshtein==0.26.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.1
Pygments==2.15.1
PyJWT==2.9.0
pyOpenSSL==23.2.0
pyparsing==3.1.4
PySocks==1.7.1
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
pyu2f==0.1.5
PyYAML==6.0.2
pyzmq==25.1.2
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.1.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.2.0
starlette==0.38.6
sympy==1.13.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.3.5
ultralytics-thop==2.0.9
urllib3==2.2.3
uvicorn==0.31.0
wcwidth==0.2.5
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.13.1
zipp==3.17.0
zstandard==0.23.0
